+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
-## [1.0.0](https://github.com/Okipa/laravel-medialibrary-ext/releases/tag/1.0.0)
+## [7.12.0](https://github.com/Okipa/laravel-medialibrary-ext/releases/tag/7.12.0)
 2019-08-27
-- First release
+- Added automatic image file name versioning for cache busting when `config('medialibrary.image_name_versioning')` is set to true.
+- Fixed missing translations loading in service provider.
+- Implemented `spatie/laravel-medialibrary:7.10.1` release.
+
+## [7.11.0](https://github.com/Okipa/laravel-medialibrary-ext/releases/tag/7.11.0)
+2019-08-27
+- First extension release.